+++ src/Ethereum/ABI.h
@@ -14,3 +14,4 @@
 #include "ABI/ParamAddress.h"
 #include "ABI/Function.h"
 #include "ABI/ParamFactory.h"
+#include "ABI/ParamStruct.h"
